added updateEvent in typeDefs and resolvers

+++ server/schemas/resolvers.js
@@ -75,6 +75,16 @@ const resolvers = {
 
       return eventQ;
     },
+    updateEvent: async (parent, {eventId, title, organizer, locationName, locationAddress, locationZipCode, description, eventDate, eventTime}, context) => {
+      if (context.user) {
+        return await User.findByIdAndUpdate({ _id: eventId }, {title, organizer, locationName, locationAddress, locationZipCode, description, eventDate, eventTime},
+           {
+          new: true,
+        });
+      }
+
+      throw new AuthenticationError("Users can only update their own events");
+    },
     addComment: async (parent, { eventId, commentText, commentAuthor }) => {
       return Event.findOneAndUpdate(
         { _id: eventId },
@@ -87,10 +97,14 @@ const resolvers = {
         }
       );
     },
-    deleteEvent: async (parent, { eventId }) => {
+    deleteEvent: async (parent, { eventId }, context) => {
+      if (context.user) {
       return Event.findOneAndDelete({ _id: eventId });
+      }
+
+      throw new AuthenticationError("Users can only delete their own events");
     },
-    removeComment: async (parent, { eventId, commentId }) => {
+    removeComment: async (parent, { eventId, commentId }, context) => {
       return Event.findOneAndUpdate(
         { _id: eventId },
         { $pull: { comments: { _id: commentId } } },
